Changes according to MR comments
